src/pyproject.toml
[tool.poetry]
name = "metaqs-api"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
gunicorn = "^20.1.0"
uvicorn = "^0.15.0"
elasticsearch = { extras = ["async"], version = "^7.14.1" }
elasticsearch-dsl = "^7.4.0"
SQLAlchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
python-slugify = "^5.0.2"
starlette-context = "^0.3.3"
fastapi-utils = "^0.2.1"
python-dotenv = "^0.20.0"
httpx = "^0.23.0"
glom = "^22.1.0"
aiocron = "^1.8"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
Sphinx = "^4.5.0"
sphinxcontrib-confluencebuilder = "^1.8.0"
pytest-cov = "^3.0.0"
sphinxcontrib-mermaid = "^0.7.1"
sphinx-serve = "^1.0.1"
sphinx-markdown-builder = "^0.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
